fix: revert to push to array but with a check first
It seems like the original logic is mutating the returned array from
underlying functions by pushing to it. And changing that to a concat
is failing the tests. So, reverting the logic back to the push command
but checking that it is an array first before spreading the errors.
If is not an array it is just pushed.